[deploy] docs
This adds some actual documentation for deploy, which is probably useful
since we told everyone it was experimentally available so they will
probably be looking at what the heck it is.

It also wires up various compoenents of the OSS build to actually work
when used from an external project.

ghstack-source-id: 5416ca59d270c488957d5527d5d20314729d0069
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69251